/test: consolidate cloud-relay store tests, parity passes (docs/CLI/iOS)
- syncCloudRelayStore tests consolidated into their colocated file
  (moved store/url/signature describes out of syncTunnelClientService.test.ts,
  deduped 3 overlapping cases); added marker-persistence-through-setRelayUrl
  case.
- iOS: testSyncMergedRelayCandidatesFoldsAdvertisedFreshestFirst pins the
  relay-fold contract (dedup, cap 3, freshest first, nil/invalid advertised
  preserves existing).
- Docs: sync README/push-notifications/ios-companion refined for the
  enabledSetByUser migration, LiveActivityIntent flow + foreground drain,
  and subset-targeted badge sync; ade-cli README relay lines updated.
